requirements.txt
streamlit>=1.30
python-dotenv>=1.0

langchain>=0.2
langgraph>=0.2
langsmith>=0.1

langchain-tavily>=0.2
langchain-huggingface>=1.0
